+++ server/src/main/java/org/allaymc/server/network/processor/PacketProcessorHolder.java
@@ -143,6 +143,7 @@ public final class PacketProcessorHolder {
         registerProcessor(ClientState.IN_GAME, new RequestAbilityPacketProcessor());
         registerProcessor(ClientState.IN_GAME, new NPCRequestPacketProcessor());
         registerProcessor(ClientState.IN_GAME, new LecternUpdatePacketProcessor());
+        registerProcessor(ClientState.IN_GAME, new CommandBlockUpdatePacketProcessor());
     }
 
     @SuppressWarnings("unchecked")